chore: extract signing function
