+++ packages/shared/src/lib/core/wallet/actions/index.ts
@@ -9,6 +9,7 @@ export * from './mintNft'
 export * from './rejectActivity'
 export * from './sendOutput'
 export * from './updateL2BalanceWithoutActivity'
+export * from './signEvmTransaction'
 export * from './signMessage'
 
 export * from './send'
